backend/requirements.txt
Django==3.2.3
django-filter==2.4.0
djangorestframework==3.12.4
django-cors-headers==4.3.1
djoser==2.1.0
gunicorn==20.1.0
psycopg2-binary==2.9.9
Pillow==9.5.0
python-dotenv==1.0.1
PyYAML==6.0
reportlab==4.4.1
drf-extra-fields==3.4.0